backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.1
daphne>=4.0
PyPDF2>=3.0
python-docx>=1.1
numpy>=1.24
matplotlib>=3.7
tiktoken>=0.5
djangorestframework-simplejwt>=5.3
